move the examples outside packages and use the deps as workspaces in examples
