+++ examples/excalidraw/with-nextjs-approuter/src/app/page.tsx
@@ -1,0 +1,10 @@
+import ExcalidrawWithClientOnly from "./excalidraw";
+
+export default function Page() {
+  return (
+    <>
+      <h1 style={{ textAlign: "center" }}>Excalidraw With Next JS</h1>
+      <ExcalidrawWithClientOnly />
+    </>
+  );
+}
