+++ packages/excalidraw/examples/with-nextjs-approuter/package.json
@@ -9,7 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@excalidraw/excalidraw": "0.17.1-7441-4e2c539",
+    "@excalidraw/excalidraw": "workspace:^",
     "next": "14.0.4",
     "react": "^18",
     "react-dom": "^18"
