--- packages/excalidraw/examples/with-script-in-browser/MobileFooter.tsx
@@ -1,20 +1,0 @@
-import type { ExcalidrawImperativeAPI } from "../../types";
-import CustomFooter from "./CustomFooter";
-const { useDevice, Footer } = window.ExcalidrawLib;
-
-const MobileFooter = ({
-  excalidrawAPI,
-}: {
-  excalidrawAPI: ExcalidrawImperativeAPI;
-}) => {
-  const device = useDevice();
-  if (device.editor.isMobile) {
-    return (
-      <Footer>
-        <CustomFooter excalidrawAPI={excalidrawAPI} />
-      </Footer>
-    );
-  }
-  return null;
-};
-export default MobileFooter;
