--- packages/excalidraw/examples/with-script-in-browser/public/index.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta
-      name="viewport"
-      content="width=device-width, initial-scale=1, shrink-to-fit=no"
-    />
-    <meta name="theme-color" content="#000000" />
-
-    <title>React App</title>
-    <script>
-      window.name = "codesandbox";
-    </script>
-    <link rel="stylesheet" href="/dist/browser/dev/index.css" />
-    <link rel="stylesheet" href="bundle.css" />
-  </head>
-
-  <body>
-    <noscript> You need to enable JavaScript to run this app. </noscript>
-    <div id="root"></div>
-    <script src="https://unpkg.com/react@18.2.0/umd/react.development.js"></script>
-    <script src="https://unpkg.com/react-dom@18.2.0/umd/react-dom.development.js"></script>
-    <!-- This is so that we use the bundled excalidraw.development.js file instead
-    of the actual source code -->
-    <script type="module">
-      import * as ExcalidrawLib from "/dist/browser/dev/index.js";
-      window.ExcalidrawLib = ExcalidrawLib;
-    </script>
-    <script type="module" src="bundle.js"></script>
-  </body>
-</html>
